[MNT] make `pyproject.toml` `pip` compatible (#383)
The current `pyproject.toml` can only be read by `poetry`, not by `pip`.

This is a serious limitation to users and developers, since it forces
them to use `poetry`.

This PR moves `pyproject.toml` to the most widely interoperable format.
This PR does an 1:1 translation, except for removing python 3.8 (which
is no longer supported) and adding python 3.13 to the range.

Also:
* updates the docs with references to `pip`
* lints impacted files