[BUG] fix missing datasets and dataset loaders crashing; also impacts quickstart tutorial (#401)
### Problem
The `wage.csv` file is missing from the package, causing the following
error when running the tutorial:

```
[Errno 2] No such file or directory: 'c:\\anaconda3\\App\\envs\\wk.pyGAM-FirstSteps\\Lib\\site-packages\\pygam\\datasets/wage.csv'
```

### Solution:
I added code that downloads the CSV from the original GitHub repository
and constructs the expected training set (X, y).

### Suggestion:
It might be simpler and more user-friendly to include wage.csv back in
the package.

---------

Co-authored-by: Franz Király <fkiraly@gcos.ai>
Co-authored-by: Thomas Kraxner <thomas.kraxner@illwerkevkw.at>