Fix error with data ownership in cholesky (#288)
Require that the fill-reducing permutation P of Cholesky factorizer owns
its data before use. Fixes #271

This sets the `OWNDATA` flag of its data to `True` - in traceback it can
be seen that `col` (line 102 of `scipy/sparse/_index.py`), which is
derived from `p`, has this flag set to false. This creates a copy of the
data if necessary (which apparently it wasn't in previous versions?)

Someone should give this patch a shot on their machine as well.